restructure app load flows to improve mobile perf
