+++ src/components/loading.tsx
@@ -5,17 +5,15 @@ import { ReactComponent as Santa } from "../assets/santa.svg";
 
 const fade = keyframes`
   50% { opacity: 1; }
-  from { transform: rotate(0deg); }
-    to { transform: rotate(360deg); }
 `;
 
 const LoaderContainer = styled.div`
   position: absolute;
   top: 40%;
-  margin-left: calc(50% - 50px);;
+  margin-left: calc(50% - 50px);
   svg {
     width: 100px;
-    animation: ${fade} 2s ease-in-out forwards infinite;
+    animation: ${fade} 0.85s ease-in-out forwards infinite;
     opacity: 0;
   }
 `;
